REFACTORING one price helper is enough
